+++ qcal/_vendor/pyrpe/__init__.py
@@ -1,0 +1,9 @@
+# This project vendors code from sandialabs/pyRPE
+# Source: https://github.com/sandialabs/pyRPE
+# License: Apache-2.0 or BSD-3-Clause
+# Vendored from commit: e09be5a
+# Files vendored under: qcal/_vendor/pyrpe/
+# Copyright 2020 National Technology & Engineering Solutions of Sandia, LLC (NTESS).
+# Under the terms of Contract DE-NA0003525 with NTESS, the U.S. Government retains
+# certain rights in this software.
+from .classical import *